DRILL-5735: UI options grouping and filtering & Metrics hints Allow search/sort in the Options webUI closes #1279
Additional details, like the description is provided as well in a JScript lookup map. This helps reduce the need for the server to constantly recreate the entire page with the description details, as the client browser can fill in these details.
Developers will be expected to update the description as old/new options are introduced or deprecated.
System table options_val also carries the description, albeit a 'shorter' version.